editoast: split the timetable import into its own module
